requirements.txt
requests>=2.28.0    
polygon-api-client>=1.25
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
tensortrade>=1.0.4
ray[rllib]==2.9.3
gymnasium==0.29.1
python-dotenv>=1.0